Use Strings in structs for test and reuse them
