+++ application/trading_service/tests/buy_eth_btc_routes.rs
@@ -1,6 +1,5 @@
 extern crate bitcoin_htlc;
 extern crate bitcoin_support;
-extern crate common_types;
 extern crate ethereum_support;
 extern crate event_store;
 extern crate rocket;
@@ -12,63 +11,14 @@ extern crate serde_json;
 extern crate trading_service;
 extern crate uuid;
 
+mod common;
+
 use bitcoin_support::{BitcoinQuantity, Network};
-use common_types::{
-    ledger::{bitcoin::Bitcoin, ethereum::Ethereum, Ledger},
-    TradingSymbol,
-};
-use ethereum_support::EthereumQuantity;
+use common::{OfferResponseBody, RedeemDetails, RequestToFund};
 use event_store::InMemoryEventStore;
-use rocket::{http::*, request::FromParam};
-use std::{fmt, sync::Arc};
+use rocket::http::*;
+use std::sync::Arc;
 use trading_service::{exchange_api_client::FakeApiClient, rocket_factory::create_rocket_instance};
-use uuid::Uuid;
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct RequestToFund {
-    address_to_fund: bitcoin_support::Address,
-    btc_amount: BitcoinQuantity,
-    eth_amount: EthereumQuantity,
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-pub struct RedeemDetails {
-    address: ethereum_support::Address,
-    data: bitcoin_htlc::secret::Secret,
-    gas: u64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub struct TradeId(Uuid); //todo extract common types from tests into own module
-
-impl From<Uuid> for TradeId {
-    fn from(uuid: Uuid) -> Self {
-        TradeId(uuid)
-    }
-}
-
-impl fmt::Display for TradeId {
-    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        self.0.fmt(f)
-    }
-}
-
-impl<'a> FromParam<'a> for TradeId {
-    type Error = uuid::ParseError;
-
-    fn from_param(param: &RawStr) -> Result<Self, <Self as FromParam>::Error> {
-        Uuid::parse_str(param.as_str()).map(|uid| TradeId::from(uid))
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct OfferResponseBody<Buy: Ledger, Sell: Ledger> {
-    pub uid: TradeId,
-    pub symbol: TradingSymbol,
-    pub rate: f64,
-    pub buy_amount: Buy::Quantity,
-    pub sell_amount: Sell::Quantity,
-}
 
 // Secret: 12345678901234567890123456789012
 // Secret hash: 51a488e06e9c69c555b8ad5e2c4629bb3135b96accd1f23451af75e06d3aee9c
@@ -101,13 +51,11 @@ fn happy_path_buy_x_eth_for_btc() {
     let mut response = request.dispatch();
 
     assert_eq!(response.status(), Status::Ok);
-    let offer_response = serde_json::from_str::<OfferResponseBody<Ethereum, Bitcoin>>(
-        &response.body_string().unwrap(),
-    ).unwrap();
+    let offer_response =
+        serde_json::from_str::<OfferResponseBody>(&response.body_string().unwrap()).unwrap();
 
     assert_eq!(
-        offer_response.symbol,
-        TradingSymbol::ETH_BTC,
+        offer_response.symbol, "ETH-BTC",
         "offer_response has correct symbol"
     );
     let uid = offer_response.uid;
@@ -125,12 +73,7 @@ fn happy_path_buy_x_eth_for_btc() {
     let funding_request =
         serde_json::from_str::<RequestToFund>(&response.body_string().unwrap()).unwrap();
 
-    assert!(
-        funding_request
-            .address_to_fund
-            .to_string()
-            .starts_with("tb1")
-    );
+    assert!(funding_request.address_to_fund.starts_with("tb1"));
 
     let request = client
         .post(format!(
